requirements.txt
evdev
dbus-python
notify2
